+++ packages/liveblocks-server/test/lib/UniqueMap.test.ts
@@ -15,6 +15,8 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
+import { describe, expect, test } from "vitest";
+
 import { UniqueMap } from "~/lib/UniqueMap";
 
 function keyFn(name: string): string {